De-duplicate TestCollisions
