+++ config/symbols.us.stcen.txt
@@ -9,6 +9,14 @@ g_eInitGeneric2 = 0x80180458;
 g_eDamageDisplayInit = 0x80180464;
 g_eRedDoorTiles = 0x801804A4;
 UNK_Invincibility0 = 0x801807B8;
+g_testCollEnemyLookup = 0x801807E4;
+g_testCollLuckCutoff = 0x80180B04;
+g_testColluCoords = 0x80180B08;
+g_testCollvCoords = 0x80180B10;
+g_testCollElementLookup = 0x80180B18;
+g_testColliFrames = 0x80180B2C;
+g_testCollPrizeTable = 0x80180B38;
+g_testCollRandTable = 0x80180B78;
 g_eDamageDisplayClut = 0x80180B94;
 g_eRedDoorUV = 0x80180BA4;
 c_GoldPrizes = 0x80180F58;
@@ -25,17 +33,14 @@ EntityMaria = 0x80190374;
 EntityElevatorStationary = 0x80190540;
 EntityUnkId1B = 0x80190A78;
 EntityMovingElevator = 0x80190B64;
-Random = 0x80190E4C;
 Update = 0x80190E7C;
 UpdateStageEntities = 0x80191178;
 TestCollisions = 0x80191280;
 EntityDamageDisplay = 0x80192398;
 InitRoomEntities = 0x80193298;
 UpdateRoomPosition = 0x80193410;
-CreateEntityFromEntity = 0x80193538;
 EntityRedDoor = 0x8019362C;
 DestroyEntity = 0x80194264;
-PreventEntityFromRespawning = 0x8019434C;
 AnimateEntity = 0x80194394;
 GetSideToPlayer = 0x801945D4;
 MoveEntity = 0x80194618;
