+++ config/symbols.us.stdre.txt
@@ -16,6 +16,14 @@ g_eBreakableanimSets = 0x80180640;
 g_eBreakableDrawModes = 0x80180650;
 g_CloneShootOrder = 0x801807D4;
 UNK_Invincibility0 = 0x8018097C;
+g_testCollEnemyLookup = 0x801809A8;
+g_testCollLuckCutoff = 0x80180CC8;
+g_testColluCoords = 0x80180CCC;
+g_testCollvCoords = 0x80180CD4;
+g_testCollElementLookup = 0x80180CDC;
+g_testColliFrames = 0x80180CF0;
+g_testCollPrizeTable = 0x80180CFC;
+g_testCollRandTable = 0x80180D3C;
 g_eDamageDisplayClut = 0x80180D58;
 g_eRedDoorUV = 0x80180D68;
 c_GoldPrizes = 0x8018111C;
@@ -37,7 +45,6 @@ EntityUnkId21 = 0x80196238;
 EntityUnkId23 = 0x80196678;
 EntityFadeToWhite1 = 0x8019697C;
 EntityFadeToWhite2 = 0x80196CC8;
-Random = 0x80196F90;
 Update = 0x80196FC0;
 UpdateStageEntities = 0x801972BC;
 TestCollisions = 0x801973C4;
@@ -45,10 +52,8 @@ EntityDamageDisplay = 0x801984DC;
 InitRoomEntities = 0x801993DC;
 UpdateRoomPosition = 0x80199554;
 CreateEntityFromCurrentEntity = 0x80199608;
-CreateEntityFromEntity = 0x8019967C;
 EntityRedDoor = 0x80199770;
 DestroyEntity = 0x8019A3A8;
-PreventEntityFromRespawning = 0x8019A490;
 AnimateEntity = 0x8019A4D8;
 MoveEntity = 0x8019A75C;
 FallEntity = 0x8019A78C;
@@ -74,7 +79,6 @@ EntityStageNamePopup = 0x801A0734;
 EntitySoulStealOrb = 0x801A13B8;
 EntityEnemyBlood = 0x801A174C;
 EntityRoomForeground = 0x801A1BFC;
-BottomCornerText = 0x801A1CE8;
 EntityUnkId17 = 0x801A2E20;
 Entity3DBackgroundHouse = 0x801A2F10;
 g_LayoutObjHorizontal = 0x801A32C4;
