+++ config/symbols.us.stno3.txt
@@ -16,6 +16,14 @@ g_eBreakableanimSets = 0x80180E80;
 g_eBreakableDrawModes = 0x80180E90;
 g_TrapDoorFlag = 0x8018123C;
 UNK_Invincibility0 = 0x80181CC0;
+g_testCollEnemyLookup = 0x80181CEC;
+g_testCollLuckCutoff = 0x8018200C;
+g_testColluCoords = 0x80182010;
+g_testCollvCoords = 0x80182018;
+g_testCollElementLookup = 0x80182020;
+g_testColliFrames = 0x80182034;
+g_testCollPrizeTable = 0x80182040;
+g_testCollRandTable = 0x80182080;
 g_eDamageDisplayClut = 0x8018209C;
 g_eRedDoorUV = 0x801820AC;
 c_GoldPrizes = 0x80182460;
@@ -80,7 +88,6 @@ CreateEntityFromCurrentEntity = 0x801C3EC4;
 CreateEntityFromEntity = 0x801C3F38;
 EntityRedDoor = 0x801C402C;
 DestroyEntity = 0x801C4C64;
-PreventEntityFromRespawning = 0x801C4D4C;
 AnimateEntity = 0x801C4D94;
 GetDistanceToPlayerX = 0x801C4F64;
 GetDistanceToPlayerY = 0x801C4FA0;
@@ -115,7 +122,6 @@ EntityStrongWargDeathBeams = 0x801CEF88;
 EntityWarg = 0x801CF740;
 EntityUnkId4B = 0x801D04A4;
 EntityWargExplosionPuffTransparent = 0x801D0624;
-BottomCornerText = 0x801D06FC;
 EntityAlucardWaterEffect = 0x801D0B34;
 EntitySplashWater = 0x801D16B4;
 EntitySurfacingWater = 0x801D1FC0;
