+++ config/symbols.us.stnp3.txt
@@ -16,6 +16,14 @@ g_eBreakableanimSets = 0x80180E58;
 g_eBreakableDrawModes = 0x80180E68;
 g_TrapDoorFlag = 0x80181104;
 UNK_Invincibility0 = 0x8018164C;
+g_testCollEnemyLookup = 0x80181678;
+g_testCollLuckCutoff = 0x80181998;
+g_testColluCoords = 0x8018199C;
+g_testCollvCoords = 0x801819A4;
+g_testCollElementLookup = 0x801819AC;
+g_testColliFrames = 0x801819C0;
+g_testCollPrizeTable = 0x801819CC;
+g_testCollRandTable = 0x80181A0C;
 g_eDamageDisplayClut = 0x80181A28;
 g_eRedDoorUV = 0x80181A38;
 c_GoldPrizes = 0x80181DEC;
