+++ config/symbols.us.stnz0.txt
@@ -16,6 +16,14 @@ g_BossFlag = 0x80181010;
 g_ElevatorTarget = 0x801813A4;
 g_ElevatorTargetPos = 0x801813A8;
 UNK_Invincibility0 = 0x80181574;
+g_testCollEnemyLookup = 0x801815A0;
+g_testCollLuckCutoff = 0x801818C0;
+g_testColluCoords = 0x801818C4;
+g_testCollvCoords = 0x801818CC;
+g_testCollElementLookup = 0x801818D4;
+g_testColliFrames = 0x801818E8;
+g_testCollPrizeTable = 0x801818F4;
+g_testCollRandTable = 0x80181934;
 g_eDamageDisplayClut = 0x80181950;
 g_eRedDoorUV = 0x80181960;
 c_GoldPrizes = 0x80181D14;
@@ -56,7 +64,6 @@ CreateEntityFromCurrentEntity = 0x801BBB4C;
 CreateEntityFromEntity = 0x801BBBC0;
 EntityRedDoor = 0x801BBCB4;
 DestroyEntity = 0x801BC8EC;
-PreventEntityFromRespawning = 0x801BC9D4;
 AnimateEntity = 0x801BCA1C;
 GetSideToPlayer = 0x801BCC5C;
 MoveEntity = 0x801BCCA0;
@@ -76,7 +83,6 @@ EntityIntenseExplosion = 0x801C15B4;
 EntitySoulStealOrb = 0x801C2AA8;
 EntityEnemyBlood = 0x801C2E3C;
 EntityRoomForeground = 0x801C32EC;
-BottomCornerText = 0x801C33D8;
 EntityBoneScimitar = 0x801C37E4;
 EntityBoneScimitarParts = 0x801C3E94;
 EntityAxeKnight = 0x801C45BC;
