+++ config/symbols.us.strwrp.txt
@@ -6,6 +6,14 @@ g_InitializeData0 = 0x8018044C;
 g_InitializeEntityData0 = 0x80180458;
 g_EInitGeneric = 0x80180470;
 UNK_Invincibility0 = 0x80180690;
+g_testCollEnemyLookup = 0x801806BC;
+g_testCollLuckCutoff = 0x801809DC;
+g_testColluCoords = 0x801809E0;
+g_testCollvCoords = 0x801809E8;
+g_testCollElementLookup = 0x801809F0;
+g_testColliFrames = 0x80180A04;
+g_testCollPrizeTable = 0x80180A10;
+g_testCollRandTable = 0x80180A50;
 c_GoldPrizes = 0x80180E30;
 c_HeartPrizes = 0x80180EC0;
 g_Rooms = 0x801811AC;
@@ -15,12 +23,12 @@ Random = 0x8018A168;
 Update = 0x8018A198;
 UpdateStageEntities = 0x8018A494;
 TestCollisions = 0x8018A59C;
+EntityDamageDisplay = 0x8018B6B4;
 InitRoomEntities = 0x8018C5B4;
 UpdateRoomPosition = 0x8018C72C;
 CreateEntityFromEntity = 0x8018C854;
 EntityIsNearPlayer = 0x8018C8D0;
 DestroyEntity = 0x8018D580;
-PreventEntityFromRespawning = 0x8018D668;
 AnimateEntity = 0x8018D6B0;
 GetDistanceToPlayerX = 0x8018D880;
 GetSideToPlayer = 0x8018D8F0;
