+++ config/symbols.us.stwrp.txt
@@ -46,18 +46,10 @@ D_80186F08 = 0x80186F08;
 g_UnkPrimHelperRot = 0x80186FC8;
 EntityWarpRoom = 0x801873A0;
 EntityWarpSmallRocks = 0x80187F1C;
-Random = 0x801881E8;
-TestCollisions = 0x8018861C;
-EntityDamageDisplay = 0x80189734;
-CreateEntityFromEntity = 0x8018A8D4;
 EntityRedDoor = 0x8018A9C8;
-DestroyEntity = 0x8018B600;
-PreventEntityFromRespawning = 0x8018B6E8;
-AllocEntity = 0x8018BE70;
 CollectGold = 0x8018CD10;
 EntityEquipItemDrop = 0x8018DA34;
 EntityStageNamePopup = 0x8019198C;
-BottomCornerText = 0x80192F40;
 g_LayoutObjHorizontal = 0x80193AB0;
 g_LayoutObjVertical = 0x80193AB4;
 g_LayoutObjPosHorizontal = 0x80193AB8;
