+++ src/st/no3/41C80.c
@@ -1,5 +1,5 @@
 #include "no3.h"
-INCLUDE_ASM("st/no3/nonmatchings/41C80", TestCollisions);
+#include "../test_collisions.h"
 
 #include "../entity_damage_display.h"
