+++ src/st/wrp/6FD0.c
@@ -481,7 +481,7 @@ u16 UNK_Invincibility0[] = {
     /* 6BA */ 0x8165,
 };
 
-u16 D_801806BC[] = {
+u16 g_testCollEnemyLookup[] = {
     0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x001E, 0x0000, 0x0000,
     0x002B, 0x0000, 0x0010, 0x0000, 0x000D, 0x0068, 0x0068, 0x0016, 0x0000,
     0x0000, 0x0000, 0x003E, 0x0000, 0x0023, 0x0050, 0x0000, 0x0000, 0x0000,
@@ -529,14 +529,14 @@ u16 D_801806BC[] = {
     0x008F, 0x0000, 0x0000, 0x0000,
 };
 
-u8 D_801809DC[] = {
+u8 g_testCollLuckCutoff[] = {
     /* 9DC */ 0x00,
     /* 9DD */ 0x40,
     /* 9DE */ 0x20,
     /* 9DF */ 0x10,
 };
 
-u8 D_801809E0[] = {
+u8 g_testColluCoords[] = {
     /* 9E0 */ 0x80,
     /* 9E1 */ 0x80,
     /* 9E2 */ 0xA0,
@@ -547,7 +547,7 @@ u8 D_801809E0[] = {
     /* 9E7 */ 0x00,
 };
 
-u8 D_801809E8[] = {
+u8 g_testCollvCoords[] = {
     /* 9E8 */ 0x60,
     /* 9E9 */ 0x60,
     /* 9EA */ 0x60,
@@ -558,7 +558,7 @@ u8 D_801809E8[] = {
     /* 9EF */ 0x00,
 };
 
-u16 D_801809F0[] = {
+u16 g_testCollElementLookup[] = {
     /* 9F0 */ 0x8000,
     /* 9F2 */ 0x4000,
     /* 9F4 */ 0x2000,
@@ -571,7 +571,7 @@ u16 D_801809F0[] = {
     /* A02 */ 0x0040,
 };
 
-u8 D_80180A04[] = {
+u8 g_testColliFrames[] = {
     /* A04 */ 0x02,
     /* A05 */ 0x04,
     /* A06 */ 0x03,
@@ -586,7 +586,7 @@ u8 D_80180A04[] = {
     /* A0F */ 0x00,
 };
 
-u16 D_80180A10[] = {
+u16 g_testCollPrizeTable[] = {
     /* A10 */ 0x0003,
     /* A12 */ 0x0000,
     /* A14 */ 0x0002,
@@ -621,7 +621,7 @@ u16 D_80180A10[] = {
     /* A4E */ 0x00C6,
 };
 
-u16 D_80180A50[] = {
+u16 g_testCollRandTable[] = {
     /* A50 */ 0x0C00,
     /* A52 */ 0x0168,
     /* A54 */ 0x0F00,
